Refactor imports for user ease Update readme Needs test rerun
