+++ galaxy_datasets/shared/gz2.py
@@ -5,10 +5,10 @@ import pandas as pd
 # TODO could eventually refactor this out of Zoobot as well
 from zoobot.shared import label_metadata
 
-from galaxy_datasets.prepared_datasets import download_utils
+from galaxy_datasets.shared import download_utils
 
 
-def setup(root, train, download):
+def gz2(root, train, download):
     resources = [
         ('https://dl.dropboxusercontent.com/s/vu77e3sh2s5c250/gz2_train_catalog.parquet', 'f489c9ec7dcf8d99f728bd00ee00b1d0'),  # the train catalog
         ('https://dl.dropboxusercontent.com/s/t8eh6f3oupndpl3/gz2_test_catalog.parquet', '8b2d74c93d35f86cc34f1d058b3b220b'),  # the test catalog
