+++ setup.py
@@ -4,7 +4,7 @@ with open("README.md", "r", encoding="utf-8") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="galaxy_datasets",
+    name="galaxy-datasets",
     version="0.0.2",
     author="Mike Walmsley",
     author_email="walmsleymk1@gmail.com",
@@ -20,5 +20,22 @@ setuptools.setup(
         "Environment :: GPU :: NVIDIA CUDA"
     ],
     packages=setuptools.find_packages(),
-    python_requires=">=3.7"
+    python_requires=">=3.7",
+    extras_require={
+        # these match zoobot's reqs
+        # TODO since zoobot depends on this, can now remove from zoobot reqs if desired?
+        'pytorch': [
+            'torch == 1.10.1',
+            'torchvision == 0.11.2',
+            'torchaudio == 0.10.1',
+            'pytorch-lightning==1.6.5',  # 1.7 requires protobuf version incompatible with tensorflow/tensorboard. Otherwise works.
+            'simplejpeg',
+            'albumentations'
+        ],
+        'tensorflow': [
+            'tensorflow >= 2.10.0',
+            'protobuf <= 3.19'  # tensorflow incompatible above this (usually resolved by pip automatically)
+        ]
+        # if you need both, specify both
+    }
 )
